make GenerateHensonNavigatorTask cacheable
Add releasing instructions
