+++ henson-plugin/src/main/java/dart/henson/plugin/internal/TaskManager.java
@@ -17,38 +17,14 @@
 
 package dart.henson.plugin.internal;
 
-import static com.android.build.gradle.internal.publishing.AndroidArtifacts.ARTIFACT_TYPE;
 import static dart.henson.plugin.util.StringUtil.capitalize;
-import static java.util.Collections.singletonList;
 
 import com.android.build.gradle.api.BaseVariant;
-import com.android.build.gradle.internal.publishing.AndroidArtifacts;
-import com.google.common.collect.Streams;
 import dart.henson.plugin.generator.HensonNavigatorGenerator;
 import java.io.File;
-import java.io.IOException;
-import java.nio.file.Files;
-import java.util.ArrayList;
-import java.util.Collections;
-import java.util.HashSet;
-import java.util.List;
-import java.util.Set;
-import java.util.zip.ZipEntry;
-import java.util.zip.ZipFile;
-import org.gradle.api.Action;
-import org.gradle.api.DefaultTask;
 import org.gradle.api.Project;
 import org.gradle.api.Task;
-import org.gradle.api.artifacts.Configuration;
-import org.gradle.api.attributes.AttributeContainer;
-import org.gradle.api.file.FileCollection;
-import org.gradle.api.internal.file.UnionFileCollection;
 import org.gradle.api.logging.Logger;
-import org.gradle.api.tasks.Input;
-import org.gradle.api.tasks.InputFiles;
-import org.gradle.api.tasks.OutputDirectory;
-import org.gradle.api.tasks.TaskAction;
-import org.gradle.api.tasks.compile.JavaCompile;
 
 public class TaskManager {
 
@@ -98,102 +74,5 @@ public class TaskManager {
     generateHensonNavigatorTask.setDependsOn(variant.getJavaCompiler().getDependsOn());
     variant.getJavaCompiler().dependsOn(generateHensonNavigatorTask);
     return generateHensonNavigatorTask;
-  }
-
-  public static class GenerateHensonNavigatorTask extends DefaultTask {
-    @Input
-    Configuration getVariantCompileClasspath() {
-      return variant.getCompileConfiguration();
-    }
-
-    @InputFiles
-    FileCollection getJarDependencies() {
-      //Thanks to Xavier Durcrohet for this
-      //https://android.googlesource.com/platform/tools/base/+/gradle_3.0.0/build-system/gradle-core/src/main/java/com/android/build/gradle/internal/scope/VariantScopeImpl.java#1037
-      Action<AttributeContainer> attributes =
-          container ->
-              container.attribute(ARTIFACT_TYPE, AndroidArtifacts.ArtifactType.CLASSES.getType());
-      boolean lenientMode = false;
-      return getVariantCompileClasspath()
-          .getIncoming()
-          .artifactView(
-              config -> {
-                config.attributes(attributes);
-                config.lenient(lenientMode);
-              })
-          .getArtifacts()
-          .getArtifactFiles();
-    }
-
-    @Input String hensonNavigatorPackageName;
-    @OutputDirectory File destinationFolder;
-    BaseVariant variant;
-    Project project;
-    Logger logger;
-    HensonNavigatorGenerator hensonNavigatorGenerator;
-
-    @TaskAction
-    public void generateHensonNavigator() {
-      JavaCompile javaCompiler = (JavaCompile) variant.getJavaCompiler();
-      FileCollection variantCompileClasspath = getJarDependencies();
-      FileCollection uft =
-          new UnionFileCollection(javaCompiler.getSource(), project.fileTree(destinationFolder));
-      javaCompiler.setSource(uft);
-      logger.debug("Analyzing configuration: " + variantCompileClasspath.getFiles());
-      Set<String> targetActivities = new HashSet<>();
-      Streams.stream(variantCompileClasspath)
-          .forEach(
-              dependency -> {
-                logger.debug("Detected dependency: {}", dependency.getAbsolutePath());
-                if (dependency.getName().endsWith(".jar")) {
-                  logger.debug("Detected navigation API dependency: {}", dependency.getName());
-                  if (!dependency.exists()) {
-                    logger.debug("Dependency jar doesn't exist {}", dependency.getAbsolutePath());
-                  } else {
-                    File file = dependency.getAbsoluteFile();
-                    List<String> entries = getJarContent(file);
-                    entries.forEach(
-                        entry -> {
-                          if (entry.matches(".*__IntentBuilder.class")) {
-                            logger.debug("Detected intent builder: {}", entry);
-                            String targetActivityFQN =
-                                entry
-                                    .substring(0, entry.length() - "__IntentBuilder.class".length())
-                                    .replace('/', '.');
-                            targetActivities.add(targetActivityFQN);
-                          }
-                        });
-                  }
-                }
-              });
-      String hensonNavigator =
-          hensonNavigatorGenerator.generateHensonNavigatorClass(
-              targetActivities, hensonNavigatorPackageName);
-      destinationFolder.mkdirs();
-      String generatedFolderName = hensonNavigatorPackageName.replace('.', '/').concat("/");
-      File generatedFolder = new File(destinationFolder, generatedFolderName);
-      generatedFolder.mkdirs();
-      File generatedFile = new File(generatedFolder, "HensonNavigator.java");
-      try {
-        logger.debug("Generating Henson navigator in " + generatedFile.getAbsolutePath());
-        logger.debug(hensonNavigator);
-        Files.write(generatedFile.toPath(), singletonList(hensonNavigator));
-      } catch (IOException e) {
-        throw new RuntimeException(e);
-      }
-    }
-
-    private List<String> getJarContent(File file) {
-      final List<String> result = new ArrayList<>();
-      try {
-        if (file.getName().endsWith(".jar")) {
-          ZipFile zip = new ZipFile(file);
-          Collections.list(zip.entries()).stream().map(ZipEntry::getName).forEach(result::add);
-        }
-      } catch (IOException e) {
-        throw new RuntimeException(e);
-      }
-      return result;
-    }
   }
 }
