+++ henson-plugin/src/main/resources/build.properties
@@ -1,2 +1,2 @@
-#Fri Feb 16 10:58:30 PST 2018
+#Tue Mar 06 17:59:16 PST 2018
 dart.version=3.0.0-RC5-SNAPSHOT
